pyproject.toml
[project]
name = "arch-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=22.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]
